Moved snapping from ProgressBar to Slider to prevent snapping when setting the value programmatically.
