+++ CHANGES
@@ -1,3 +1,6 @@
+[1.9.4]
+- Moved snapping from ProgressBar to Slider to prevent snapping when setting the value programmatically.
+
 [1.9.3]
 - Switched to MobiDevelop's RoboVM fork (http://robovm.mobidevelop.com)
 - Addition of Intel Multi-OS Engine backend for deploying to iOS
