+++ gdx/src/com/badlogic/gdx/scenes/scene2d/ui/Slider.java
@@ -16,6 +16,8 @@
 
 package com.badlogic.gdx.scenes.scene2d.ui;
 
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Input.Keys;
 import com.badlogic.gdx.graphics.g2d.NinePatch;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Interpolation;
@@ -39,6 +41,8 @@ public class Slider extends ProgressBar {
 	int draggingPointer = -1;
 	boolean mouseOver;
 	private Interpolation visualInterpolationInverse = Interpolation.linear;
+	private float[] snapValues;
+	private float threshold;
 
 	public Slider (float min, float max, float stepSize, boolean vertical, Skin skin) {
 		this(min, max, stepSize, vertical, skin.get("default-" + (vertical ? "vertical" : "horizontal"), SliderStyle.class));
@@ -58,8 +62,6 @@ public class Slider extends ProgressBar {
 	 * @param style the {@link SliderStyle} */
 	public Slider (float min, float max, float stepSize, boolean vertical, SliderStyle style) {
 		super(min, max, stepSize, vertical, style);
-
-		shiftIgnoresSnap = true;
 
 		addListener(new InputListener() {
 			public boolean touchDown (InputEvent event, float x, float y, int pointer, int button) {
@@ -112,8 +114,8 @@ public class Slider extends ProgressBar {
 	protected Drawable getKnobDrawable () {
 		SliderStyle style = getStyle();
 		return (disabled && style.disabledKnob != null) ? style.disabledKnob
-			: (isDragging() && style.knobDown != null) ? style.knobDown : ((mouseOver && style.knobOver != null) ? style.knobOver
-				: style.knob);
+			: (isDragging() && style.knobDown != null) ? style.knobDown
+				: ((mouseOver && style.knobOver != null) ? style.knobOver : style.knob);
 	}
 
 	boolean calculatePositionAndValue (float x, float y) {
@@ -144,9 +146,26 @@ public class Slider extends ProgressBar {
 		}
 
 		float oldValue = value;
+		if (!Gdx.input.isKeyPressed(Keys.SHIFT_LEFT) && !Gdx.input.isKeyPressed(Keys.SHIFT_RIGHT)) value = snap(value);
 		boolean valueSet = setValue(value);
 		if (value == oldValue) position = oldPosition;
 		return valueSet;
+	}
+
+	/** Returns a snapped value. */
+	private float snap (float value) {
+		if (snapValues == null) return value;
+		for (int i = 0; i < snapValues.length; i++) {
+			if (Math.abs(value - snapValues[i]) <= threshold) return snapValues[i];
+		}
+		return value;
+	}
+
+	/** Will make this progress bar snap to the specified values, if the knob is within the threshold.
+	 * @param values May be null. */
+	public void setSnapToValues (float[] values, float threshold) {
+		this.snapValues = values;
+		this.threshold = threshold;
 	}
 
 	/** Returns true if the slider is being dragged. */
